remove helper used only once
