+++ ui/tests/integration/components/secret-list-header-test.js
@@ -5,11 +5,21 @@
 
 import { module, test } from 'qunit';
 import { setupRenderingTest } from 'ember-qunit';
-import { render } from '@ember/test-helpers';
+import { getContext, render } from '@ember/test-helpers';
 import hbs from 'htmlbars-inline-precompile';
 import { supportedSecretBackends } from 'vault/helpers/supported-secret-backends';
 import { setupMirage } from 'ember-cli-mirage/test-support';
-import mirageToModels from 'vault/tests/helpers/mirage-to-models';
+
+const mirageToModels = (data) => {
+  const context = getContext();
+  const store = context.owner.lookup('service:store');
+  const modelName = Array.isArray(data) ? data[0].modelName : data.modelName;
+  const json = context.server.serializerOrRegistry.serialize(data);
+  store.push(json);
+  return Array.isArray(data)
+    ? data.map(({ id }) => store.peekRecord(modelName, id))
+    : store.peekRecord(modelName, data.id);
+};
 
 module('Integration | Component | secret-list-header', function (hooks) {
   setupRenderingTest(hooks);
